backend/requirements.txt
fastapi
uvicorn
python-whois
tldextract
email-validator
dnspython
requests